requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.10.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
